Update .gitignore
This just read the file and format to capital letters, no gaps, and no new lines. I am stuck in stripping the header in order to do the reading frames because when I take the headers out all the sequences get joint as I send you in the code last week by email. Any idea how to continue?